+++ app/src/main/java/umc/study/umc_7th/album/AlbumItem.kt
@@ -1,9 +1,8 @@
-package umc.study.umc_7th
+package umc.study.umc_7th.album
 
 import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -29,6 +28,8 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import umc.study.umc_7th.R
+import umc.study.umc_7th.albumData
 
 @RequiresApi(Build.VERSION_CODES.P)
 @Composable
