+++ app/src/main/java/umc/study/umc_7th/main/home/BottomNavigationBar.kt
@@ -1,24 +1,15 @@
-package umc.study.umc_7th
+package umc.study.umc_7th.main.home
 
-import android.os.Build
-import androidx.annotation.RequiresApi
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 //import androidx.compose.foundation.layout.FlowRowScopeInstance.weight
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.Icon
-import androidx.compose.material3.NavigationBarItemDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -27,16 +18,12 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.style.LineBreak
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.res.imageResource
 import androidx.navigation.NavController
+import umc.study.umc_7th.R
 
 enum class DestinationClass(
     val route : String,
@@ -48,10 +35,14 @@ enum class DestinationClass(
         "homeFragment","홈", R.drawable.ic_bottom_home_select, R.drawable.ic_bottom_home_no_select
     ),
     Around(
-        "aroundFragment","둘러보기", R.drawable.ic_bottom_look_select, R.drawable.ic_bottom_look_no_select
+        "aroundFragment","둘러보기",
+        R.drawable.ic_bottom_look_select,
+        R.drawable.ic_bottom_look_no_select
     ),
     Search (
-        "searchFragment","검색", R.drawable.ic_bottom_search_select, R.drawable.ic_bottom_search_no_select
+        "searchFragment","검색",
+        R.drawable.ic_bottom_search_select,
+        R.drawable.ic_bottom_search_no_select
     ),
     Storage(
         "lockerFragment","보관함", R.drawable.ic_bottom_my_select, R.drawable.ic_bottom_my_no_select
