+++ app/src/main/java/umc/study/umc_7th/main/home/Promotion.kt
@@ -1,12 +1,10 @@
-package umc.study.umc_7th
+package umc.study.umc_7th.main.home
 
 import android.os.Build
-import android.provider.ContactsContract.CommonDataKinds.Im
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.ImageBitmap
@@ -14,6 +12,7 @@ import androidx.compose.ui.res.imageResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
+import umc.study.umc_7th.R
 
 // 야놀자, 플로 배너 모두 사용 가능함
 @RequiresApi(Build.VERSION_CODES.P)
