+++ app/src/main/java/umc/study/umc_7th/main/home/SnsSharebar.kt
@@ -1,4 +1,4 @@
-package umc.study.umc_7th
+package umc.study.umc_7th.main.home
 
 import android.os.Build
 import androidx.annotation.RequiresApi
@@ -17,6 +17,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import umc.study.umc_7th.R
 
 @RequiresApi(Build.VERSION_CODES.P)
 @Composable
