+++ app/src/main/java/umc/study/umc_7th/main/song/Lyric.kt
@@ -1,8 +1,7 @@
-package umc.study.umc_7th
+package umc.study.umc_7th.main.song
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Text
